+++ wireform-kafka/test/Main.hs
@@ -39,6 +39,7 @@ import qualified Client.ProducerTimeoutSpec
 import qualified Client.ProducerConsumerLifecycleSpec
 import qualified Client.TransactionSpec
 import qualified Client.TransactionCoordinatorSpec
+import qualified Client.ProducerTransactionWiringSpec
 import qualified Client.AdminClientSpec
 import qualified Client.MockBrokerSpec
 import qualified Client.MockBrokerFailureModesSpec
@@ -106,6 +107,7 @@ clientTests = testGroup "Client"
   , Client.ProducerConsumerLifecycleSpec.lifecycleSpec
   , Client.TransactionSpec.transactionSpec
   , Client.TransactionCoordinatorSpec.transactionCoordinatorSpec
+  , Client.ProducerTransactionWiringSpec.tests
   , Client.AdminClientSpec.tests
   , Client.GroupSpec.groupSpec
   , Client.MockBrokerSpec.tests